implemented paralysis counter pt. 2

+++ include/pokemon.h
@@ -246,7 +246,7 @@ struct BattleMove
     s8 priority;
     u32 flags;
     u8 split;
-    u8 argument;
+    u16 argument;
 };
 
 #define SPINDA_SPOT_WIDTH 16
